Refactor document handling and add Azure Blob support
- Changed `DocumentId` type from `int` to `long` in `Document.cs`.
- Updated `ExcelImportHelper` to include `AzureBlobStorageHelper` and modified `ImportDataFromExcel` to accept an `IFormFile`.
- Refactored Excel import logic for better clarity and efficiency.
- Changed return type of `UploadFileAsync` in `S3Helper` and updated `S3Controller` accordingly.
- Updated namespaces in `UserController.cs` for view models.
- Registered `AzureBlobStorageHelper` in `Program.cs`.
- Added Azure Blob Storage configuration to `appsettings.Development.json`.
- Created `AzureBlobStorageHelper` for file management in Azure.
- Added `BlobStorageController` for handling file operations in Azure.
- Updated project files to include Azure Storage Blobs package.

+++ SKCE.Valuation.API/Examination.Models/DbModels/Common/Document.cs
@@ -13,7 +13,7 @@ namespace SKCE.Examination.Models.DbModels.Common
     {
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public int DocumentId { get; set; }
+        public long DocumentId { get; set; }
         public required string Name { get; set; }
         public required string Url { get; set; }
     }
